ARMAspec {TSA}	R Documentation

Theoretical spectral density function of a stationary ARMA model

Description

Computes and plots the theoretical spectral density function of a stationary ARMA model

Usage

ARMAspec(model, freq = seq(0, 0.5, 0.001), plot = TRUE, ...)

Arguments

model
an arma model

freq
vector of frequency over which the spectral density is computed

plot
if true, plot the spectral density function; default is true

...
other parameters to be passed to the plot function

Value

a list:

spec
spectral density values

freq
same as freq in the input

model
the arma model

Author(s)

Kung-Sik Chan

See Also

spec

Examples

theta=.9 # Reset theta for other MA(1) plots
ARMAspec(model=list(ma=-theta)) 

[Package TSA version 1.01 Index]