BoxCox.ar {TSA}	R Documentation

Determine the power transformation for serially correlated data

Description

Determine the appropriate power transformation for time-series data. The objective is to estimate the power transformation so that the transformed time series is approximately a Gaussian AR process.

Usage

BoxCox.ar(y, order, lambda = seq(-2, 2, 0.01), plotit = TRUE, 
method = c("mle", "yule-walker", "burg", "ols", "yw"), ...)

Arguments

y
univariate time series (must be positive)

order
AR order for the data; if missing, the order is determined by AIC for the log-transformed data

lambda
a vector of candidate power transformation values; if missing, it is set to be from -2 to 2, with increment .01

plotit
logical value, if true, plot the profile log-likelihood for the power estimator

method
method of AR estimation; default is "mle"

...
other parameters to be passed to the ar function

Value

A list that contains the following:

lambda
candidate power transformation parameter values

loglike
profile log-likelihood

mle
maximum likelihood estimate of the power transformation value

ci
95% C.I. of the power transformation value

Note

The procedure is very computer intensive. Be patient for the outcome

Author(s)

Kung-Sik Chan

Examples

data(hare)
# hare.transf=BoxCox.ar(y=hare)
# hare.transf$ci

[Package TSA version 1.01 Index]