Keenan.test {TSA}	R Documentation

Keenan's one-degree test for nonlinearity

Description

Carry out Keenan's 1-degree test for nonlinearity against the null hypothesis that the time series follows some AR process.

Usage

Keenan.test(x, order, ...)

Arguments

x
time series

order
working AR order; if missing, it is estimated by minimizing AIC via the ar function.

...
user-supplied options to the ar function.

Details

The test is designed to have optimal local power against depature from the linear autoregressive function in the direction of the square of the linear autoregressive function.

Value

A list containing the following components

test.stat
The observed test statistic

p.value
p-value of the test

order
working AR order

Author(s)

Kung-Sik Chan

References

Keenan, D. M. (1985), A Tukey nonadditivity-type test for time series Nonlinearity, Biometrika, 72, 39-44.

See Also

Tsay.test,tlrt

Examples

data(spots)
Keenan.test(sqrt(spots))

[Package TSA version 1.01 Index]