Tsay.test {TSA}	R Documentation

Tsay's Test for nonlinearity

Description

Carry out Tsay's test for quadratic nonlinearity in a time series.

Usage

Tsay.test(x, order, ...)

Arguments

x
time series

order
working linear AR order; if missing, it will be estimated via the ar function by minimizing AIC

...
options to be passed to the ar function

Details

The null hypothesis is that the true model is an AR process. The AR order, if missing, is estimated by minimizing AIC via the ar function, i.e. fitting autoregressive model to the data. The default fitting method of the ar function is "yule-walker."

Value

A list containing the following components

test.stat
The observed test statistic

p.value
p-value of the test

order
working AR order

Author(s)

Kung-Sik Chan

References

Tsay, R. S. (1986), Nonlinearity test for time series, Biometrika, 73, 461-466.

See Also

Tsay.test,tlrt

Examples

data(spots)
Tsay.test(sqrt(spots))

[Package TSA version 1.01 Index]