days {TSA}	R Documentation

Number of days between payment to Winegard Corp. / time series

Description

Accounts receivable data. Number of days until a distributor of Winegard Company products pays their account.

Usage

data(days)

Format

The format is: Time-Series [1:130] from 1 to 130: 39 39 41 26 28 28 25 26 24 38 ...

Source

Personal communication from Mark Selergren, Vice President, Winegard, Inc., Burlington, Iowa.

Examples

data(days)
## maybe str(days) ; plot(days) ...

[Package TSA version 1.01 Index]