detectAO {TSA}	R Documentation

Additive Outlier Detection

Description

This function serves to detect whether there are any additive outliers (AO). It implements the test statistic lambda_{2,t} proposed by Chang, Chen and Tiao (1988).

Usage

detectAO(object, alpha = 0.05, robust = TRUE)

Arguments

object
a fitted ARIMA model

alpha
family significance level (5% is the default) Bonferroni rule is used to control the family error rate.

robust
if true, the noise standard deviation is estimated by mean absolute residuals times sqrt(pi/2). Otherwise, it is the estimated by sqrt(sigma2) from the arima fit.

Value

A list containing the following components:

ind
the time indices of potential AO

lambda2
the corresponding test statistics

Author(s)

Kung-Sik Chan

References

Chang, I.H., Tiao, G.C. and C. Chen (1988). Estimation of Time Series Parameters in the Presence of Outliers. Technometrics, 30, 193-204.

See Also

detectIO

Examples

set.seed(12345)
y=arima.sim(model=list(ar=.8,ma=.5),n.start=158,n=100)
y[10]
y[10]=10
y=ts(y,freq=1,start=1)
plot(y,type='o')
acf(y)
pacf(y)
eacf(y)
m1=arima(y,order=c(1,0,0))
m1
detectAO(m1)
detectAO(m1, robust=FALSE)
detectIO(m1)

[Package TSA version 1.01 Index]