eacf {TSA}	R Documentation

Compute the sample extended acf (ESACF)

Description

Computes the sample extended acf (ESACF) for the time series stored in z. The matrix of ESACF with the AR order up to ar.max and the MA order up to ma.max is stored in the matrix EACFM.

Usage

eacf(z, ar.max = 7, ma.max = 13)

Arguments

z
the time series data

ar.max
maximum AR order; default=7

ma.max
maximum MA order; default=13

Value

A list containing the following two components:

eacf
a matrix of sample extended ACF

symbol
corresponding matrix of symbols indicating the significance of the ESACF

Side effect of the eacf function: The function prints a coded ESACF table with significant values denoted by * and nosignificant values by 0.

Author(s)

Kung-Sik Chan

References

Tsay, R. and Tiao, G. (1984). "Consistent Estimates of Autoregressive Parameters and Extended Sample Autocorrelation Function for Stationary and Nonstationary ARMA Models." Journal of the American Statistical Association, 79 (385), pp. 84-96.

Examples

data(arma11.s)
eacf(arma11.s)

[Package TSA version 1.01 Index]