euph {TSA}	R Documentation

A digitized sound file of a B flat played on a euphonium

Description

A digitized sound file of about 0.4 seconds of a B flat just below middle C played on a euphonium by one of the authors (JDC), a member of the group Tempered Brass.

Usage

data(euph)

Format

The format is: Time-Series [1:1105] from 1 to 1105: 0.244 0.635 0.712 0.608 0.317 ...

Examples

data(euph)
## maybe str(euph) ; plot(euph) ...

[Package TSA version 1.01 Index]