garch.sim {TSA}	R Documentation

Simulate a GARCH process

Description

Simulate a GARCH process.

Usage

garch.sim(alpha, beta, n = 100, rnd = rnorm, ntrans = 100,...)

Arguments

alpha
The vector of ARCH coefficients including the intercept term as the first element

beta
The vector of GARCH coefficients

n
sample size

rnd
random number generator for the noise; default is normal

ntrans
burn-in size, i.e. number of initial simulated data to be discarded

...
parameters to be passed to the random number generator

Details

Simulate data from the GARCH(p,q) model: x_t=σ_{t|t-1} e_t where \{e_t\} is iid, e_t independent of past x_{t-s}, s=1,2,…, and

σ_{t|t-1}=∑_{j=1}^p β_j σ_{t-j|t-j-1}+ α_0+∑_{j=1}^q α_j x_{t-i}^2

Value

simulated GARCH time series of size n.

Author(s)

Kung-Sik Chan

Examples

set.seed(1235678)
garch01.sim=garch.sim(alpha=c(.01,.9),n=500)
plot(garch01.sim,type='l', main='',ylab=expression(r[t]),xlab='t')

[Package TSA version 1.01 Index]