hare {TSA}	R Documentation

Canadian hare data/ time series

Description

Annual number of hare data.

Usage

data(hare)

Format

The format is: Time-Series [1:31] from 1905 to 1935: 50 20 20 22 27 50 55 78 70 59 ...

Details

These are yearly hare abundances for the main drainage of the Hudson Bay, based on trapper questionnaires.

Source

MacLulich, D. A. (1937) Fluctuations in the Number of the Varying Hare (Lepus americanus) (Univ. of Toronto Press, Toronto)

References

Stenseth, N. C., Falck, W., Bjornstad, O. N. and Krebs. C. J. (1997) Population regulation in snowshoe hare and Canadian lynx: Asymmetric food web configurations between hare and lynx. Proc. Natl. Acad. Sci., 94, 5147-5152.

Examples

data(hare)

[Package TSA version 1.01 Index]