plot.armasubsets {TSA}	R Documentation

Plot the Best Subset ARMA models

Description

This function is adapted from the plot.regsubsets function of the leaps package, and its main use is to plot the output from the armasubsets function.

Usage

## S3 method for class 'armasubsets'
plot(x, labels = obj$xnames, main = NULL, 
scale = c("BIC", "AICc", "AIC", "Cp", "adjR2", "R2"), 
col = gray(c(seq(0.4, 0.7, length = 10), 0.9)), draw.grid = TRUE, 
axis.at.3 = TRUE, ...)

Arguments

x
an object of class armasubsets

labels
variable names

main
title for plot

scale
which summary statistic to use for ordering plots

col
the last color should be close to but distinct from white

draw.grid
a logical argument; if it is true (default), gray grid lines are superimposed on the graph.

axis.at.3
a logical argument; if if it is true (default), the x-labels are drawn on the upper horizontal axis.

...
other arguments

Value

Plot the few best subset ARMA models.

Author(s)

Kung-Sik Chan, based on previoud work by Thomas Lumley and Merlise Clyde

See Also

armasubsets

Examples

set.seed(53331)
test=arima.sim(model=list(ar=c(rep(0,11),.8),ma=c(rep(0,11),0.7)),n=120)
res=armasubsets(y=test,nar=14,nma=14,y.name='test',ar.method='ols')
plot(res)

[Package TSA version 1.01 Index]