summary.armasubsets {TSA}	R Documentation

Summary of output from the armasubsets function

Description

Add the calculation of AIC and AICc. See the help manual of regsubsets function of the leaps package

Usage

## S3 method for class 'armasubsets'
summary(object, all.best = TRUE, matrix = TRUE, matrix.logical = FALSE,
 df = NULL, ...)

Arguments

object
armasubsets object

all.best
Show all the best subsets or just one of each size

matrix
Show a matrix of the variables in each model or just summary statistics

matrix.logical
With matrix=TRUE, the matrix is logical TRUE/FALSE or string "*"/code" "

df
Specify a number of degrees of freedom for the summary statistics. The default is n-1

...
Other arguments for future methods

Author(s)

Kung-Sik Chan, based on previous work of Thomas Lumley

[Package TSA version 1.01 Index]