tar {TSA}	R Documentation

Estimation of a TAR model

Description

Estimation of a two-regime TAR model.

Usage

tar(y, p1, p2, d, is.constant1 = TRUE, is.constant2 = TRUE, transform = "no",
 center = FALSE, standard = FALSE, estimate.thd = TRUE, threshold, 
method = c("MAIC", "CLS")[1], a = 0.05, b = 0.95, order.select = TRUE, print = FALSE)

Arguments

y
time series

p1
AR order of the lower regime

p2
AR order of the upper regime

d
delay parameter

is.constant1
if True, intercept included in the lower regime, otherwise the intercept is fixed at zero

is.constant2
similar to is.constant1 but for the upper regime

transform
available transformations: "no" (i.e. use raw data), "log", "log10" and "sqrt"

center
if set to be True, data are centered before analysis

standard
if set to be True, data are standardized before analysis

estimate.thd
if True, threshold parameter is estimated, otherwise it is fixed at the value supplied by threshold

threshold
known threshold value, only needed to be supplied if estimate.thd is set to be False.

method
"MAIC": estimate the TAR model by minimizing the AIC; "CLS": estimate the TAR model by the method of Conditional Least Squares.

a
lower percent; the threshold is searched over the interval defined by the a*100 percentile to the b*100 percentile of the time-series variable

b
upper percent

order.select
If method is "MAIC", setting order.select to True will enable the function to further select the AR order in each regime by minimizing AIC

print
if True, the estimated model will be printed

Details

The two-regime Threshold Autoregressive (TAR) model is given by the following formula:

Y_t = φ_{1,0}+φ_{1,1} Y_{t-1} +…+ φ_{1,p} Y_{t-p_1} +σ_1 e_t, \mbox{ if } Y_{t-d}≤ r

Y_t = φ_{2,0}+φ_{2,1} Y_{t-1} +…+φ_{2,p_2} Y_{t-p}+σ_2 e_t, \mbox{ if } Y_{t-d} > r.

where r is the threshold and d the delay.

Value

A list of class "TAR" which can be further processed by the by the predict and tsdiag functions.

Author(s)

Kung-Sik Chan

References

Tong, H. (1990) "Non-linear Time Series, a Dynamical System Approach," Clarendon Press Oxford

"Time Series Analysis, with Applications in R" by J.D. Cryer and K.S. Chan

See Also

predict.TAR, tsdiag.TAR, tar.sim, tar.skeleton

Examples

data(prey.eq)
prey.tar.1=tar(y=log(prey.eq),p1=4,p2=4,d=3,a=.1,b=.9,print=TRUE)

[Package TSA version 1.01 Index]