tar.skeleton {TSA}	R Documentation

Find the asympotitc behavior of the skeleton of a TAR model

Description

The skeleton of a TAR model is obtained by suppressing the noise term from the TAR model.

Usage

tar.skeleton(object, Phi1, Phi2, thd, d, p, ntransient = 500, n = 500, 
xstart, plot = TRUE,n.skeleton = 50)

Arguments

object
a TAR model fitted by the tar function; if it is supplied, the model parameters and initial values are extracted from it

ntransient
the burn-in size

n
sample size of the skeleton trajectory

Phi1
the coefficient vector of the lower-regime model

Phi2
the coefficient vector of the upper-regime model

thd
threshold

d
delay

p
maximum autoregressive order

xstart
initial values for the iteration of the skeleton

plot
if True, the time series plot of the skeleton is drawn

n.skeleton
number of last n.skeleton points of the skeleton to be plotted

Details

The two-regime Threshold Autoregressive (TAR) model is given by the following formula:

Y_t = φ_{1,0}+φ_{1,1} Y_{t-1} +…+ φ_{1,p} Y_{t-p_1} +σ_1 e_t, \mbox{ if } Y_{t-d}≤ r

Y_t = φ_{2,0}+φ_{2,1} Y_{t-1} +…+φ_{2,p_2} Y_{t-p}+σ_2 e_t, \mbox{ if } Y_{t-d} > r.

where r is the threshold and d the delay.

Value

A vector that contains the trajectory of the skeleton, with the burn-in discarded.

Author(s)

Kung-Sik Chan

References

Tong, H. (1990) "Non-linear Time Series, a Dynamical System Approach," Clarendon Press Oxford. "Time Series Analysis, with Applications in R" by J.D. Cryer and K.S. Chan

See Also

tar

Examples

data(prey.eq)
prey.tar.1=tar(y=log(prey.eq),p1=4,p2=4,d=3,a=.1,b=.9,print=TRUE)
tar.skeleton(prey.tar.1)

[Package TSA version 1.01 Index]