tbone {TSA}	R Documentation

A digitized sound file of a B flat played on a tenor trombone

Description

A digitized sound file of about 0.4 seconds of a B flat just below middle C played on a tenor trombone by Chuck Kreeb, a member of Tempered Brass and a friend of one of the authors.

Usage

data(tbone)

Format

The format is: Time-Series [1:17689] from 1 to 17689: 0.0769 0.0862 0.0961 0.1050 0.1129 ...

Examples

data(tbone)
## maybe str(tbone) ; plot(tbone) ...

[Package TSA version 1.01 Index]