winnebago {TSA}	R Documentation

Monthly unit sales of recreational vehicles / time series

Description

Monthly unit sales of recreational vehicles from Winnebago, Inc., Forrest City, Iowa, from 11/1966 - 02/1972.

Usage

data(winnebago)

Format

The format is: Time-Series [1:64] from 1967 to 1972: 61 48 53 78 75 58 146 193 124 120 ...

Source

Roberts, H. V., Data Analysis for Managers with Minitab, second edition, The Scientific Press, 1991.

Examples

data(winnebago)
## maybe str(winnebago) ; plot(winnebago) ...

[Package TSA version 1.01 Index]